Moved dropdown into its own component.
